Use the correct ATLAS library on openSUSE 42 and 15
Previously, R was configured to use the reference BLAS/LAPACK libraries --
atlascpp-devel was an unrelated library. ATLAS isn't available in the official
openSUSE repos so we get it from a 3rd party repo.

NOTE: we used serial ATLAS (libsatlas) on openSUSE 15 because R found that the
threaded ATLAS (libtatlas) did not support double complex BLAS.